* Fix 'e foo' output * Fix io.ffio warnings * Fix '.' key in visual mode (sr pc) * Various 64bit related fixes * Various fixes in r2-swig   - Needs valaswig from hg (update)   - r_config and

+++ libr/io/p/io_shm.c
@@ -70,9 +70,9 @@ static int shm__open(struct r_io_t *io, const char *pathname, int flags, int mod
 		// connect
 		shm_buf= shmat(atoi(ptr), 0, 0);
 
-		if (((int)(shm_buf)) != -1) {
+		if (((int)(size_t)(shm_buf)) != -1) {
 			printf("Connected to shared memory 0x%08x\n", atoi(ptr));
-			shm_fd = (int)&shm_buf;
+			shm_fd = (int)(size_t)&shm_buf;
 		} else	{
 			printf("Cannot connect to shared memory (%d)\n", atoi(ptr));
 			shm_buf = NULL;
